requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
imbalanced-learn==0.12.3
shap==0.45.1
matplotlib==3.9.0
seaborn==0.13.2
jupyter==1.0.0
mlflow==2.13.0
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0   # ← NEW: needed for BaseSettings
python-dotenv==1.0.1
streamlit==1.35.0
plotly==5.22.0
reportlab==4.2.0
requests==2.32.3
joblib==1.4.2